This is the initial branch update from master.
In 2007 we decided that Web Design Automation was going to be a thing so we created a very fast, dynamic website generator. Since then, many solutions have been developed that mimic the functionality to one extent or another of the SoNET Web Engine.  You are very welcomed to use it to make your own solutions. So go ahead and have fun.